refactor(codegen): move shared MIR helpers out of transforms
