+++ crates/codegen/src/transform/check_elim.rs
@@ -27,7 +27,7 @@ use crate::{
     analysis::CfgInfo,
     mir::{BlockId, Function, InstKind, Terminator, Value, ValueId},
     pass::FunctionPass,
-    transform::repair_reachability_phis,
+    utils::repair_reachability_phis,
 };
 use alloy_primitives::U256;
 use solar_data_structures::map::{FxHashMap, FxHashSet};
