+++ crates/codegen/src/transform/loop_canonicalize.rs
@@ -15,7 +15,7 @@ use crate::{
     analysis::LoopAnalyzer,
     mir::{BlockId, Function, InstId, InstKind, Instruction, Terminator, Value, ValueId},
     pass::FunctionPass,
-    transform::repair_reachability_phis,
+    utils::repair_reachability_phis,
 };
 use solar_data_structures::map::FxHashSet;
