+++ crates/codegen/src/transform/sccp.rs
@@ -17,7 +17,7 @@
 use crate::{
     mir::{BlockId, Function, Immediate, InstId, InstKind, MirType, Terminator, Value, ValueId},
     pass::FunctionPass,
-    transform::{const_eval, repair_reachability_phis},
+    utils::{const_eval, repair_reachability_phis},
 };
 use alloy_primitives::U256;
 use solar_data_structures::map::{FxHashMap, FxHashSet};
